+++ mopidy/mpd/protocol/status.py
@@ -34,12 +34,12 @@ def currentsong(context):
         Displays the song info of the current song (same song that is
         identified in status).
     """
-    tl_track = context.core.playback.current_metadata_track.get()
-    if tl_track is None:
-        tl_track = context.core.playback.current_tl_track.get()
+    tl_track = context.core.playback.current_tl_track.get()
+    stream_title = context.core.playback.get_stream_title().get()
     if tl_track is not None:
         position = context.core.tracklist.index(tl_track).get()
-        return translator.track_to_mpd_format(tl_track, position=position)
+        return translator.track_to_mpd_format(
+            tl_track, position=position, stream_title=stream_title)
 
 
 @protocol.commands.add('idle', list_command=False)